% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(formula, data, digit = 3, detailed = FALSE)
}
\arguments{
\item{formula}{regression formula of interests}

\item{data}{data used for regression model}

\item{digit}{digits of output numbers (except p-value)}

\item{detailed}{whether output a detailed table}
}
\value{
the square of x
}
\description{
Gets the square of a number
}
\examples{
square(3)

}
