% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(formula, data, digit = 3, detailed = FALSE)
}
\arguments{
\item{formula}{regression formula of interests}

\item{data}{data used for regression model}

\item{digit}{digits of output numbers (except p-value)}

\item{detailed}{whether output a detailed table}
}
\value{
the fitted model information
}
\description{
linear_model function is used to fit linear models.
It can be seen as a basic version of `lm` function.
}
\examples{
linear_model(formula, data, digit = 3, detailed = FALSE)

}
